Muevo funciones de helpers a otros módulos
