+++ ckanext/gobar_theme/helpers/helpers.py
@@ -5,25 +5,6 @@ import ckan.logic as logic
 from markupsafe import Markup
 
 
-def accepted_mime_types():
-    return [
-        'html',
-        'json',
-        'xml',
-        'text',
-        'csv',
-        'xls',
-        'api',
-        'pdf',
-        'zip',
-        'rdf',
-        'nquad',
-        'ntriples',
-        'turtle',
-        'shp'
-    ]
-
-
 def package_resources(pkg_id):
     package = logic.get_action('package_show')({}, {'id': pkg_id})
     return package['resources']
@@ -51,21 +32,3 @@ def jsondump(field=''):
     return Markup(json.dumps(field))
 
 
-def delete_column_from_csv_file(csv_path, column_name):
-    with open(csv_path, 'rb') as source:
-        rdr = csv.reader(source)
-        first_row = next(rdr)
-        column_position = None
-        try:
-            column_position = first_row.index(column_name)
-        except ValueError:
-            # No existe una columna con el nombre que llegó por parámetro -> se usará el csv tal y como está
-            return
-        source.seek(0)
-        list_with_rows = []
-        for r in rdr:
-            list_with_rows.append(tuple((r[x] for x in range(len(r)) if x != column_position)))
-    with open(csv_path, 'wb') as result:
-        wtr = csv.writer(result)
-        for r in list_with_rows:
-            wtr.writerow(tuple(x for x in r))
